Extract view state class to separate file
